Address CodeQL findings without changing netplay behavior.
Use vps_ssh with host-key verification for deploy scripts, redact dashboard
deploy credentials from stdout, fix launcher log format specifiers, and add
CodeQL suppressions where behavior is already safe. Enable Dependabot for Actions.

+++ src/launcher/github_release_client.cxx
@@ -287,7 +287,7 @@ namespace launcher {
 			}
 			if (exitCode != 0) {
 				char buf[64] = { 0 };
-				snprintf(buf, sizeof(buf), "tar failed with exit code %lu", exitCode);
+				snprintf(buf, sizeof(buf), "tar failed with exit code %u", (unsigned)exitCode);
 				AppendUpdateLog(buf);
 				return false;
 			}
